pyproject.toml
[tool.poetry]
name = "dicom-test"
version = "0.1.0"
description = ""
authors = ["Anthony Sazonov <jericho9378@gmail.com>"]
readme = "README.md"
packages = [{include = "dicom_test"}]

[tool.poetry.dependencies]
python = "^3.10"
pydicom = "^2.4.3"
pynetdicom = "^2.0.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
